Removing LTSpice from all prefixes
