+++ PyLTSpice/RawRead.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 
 # -------------------------------------------------------------------------------
-# Name:        LTSpice_RawRead.py
+# Name:        RawRead.py
 # Purpose:     Process LTSpice output files and align data for usage in a spread-
 #              sheet tool such as Excel, or Calc.
 #
@@ -14,7 +14,7 @@
 
 """
 This module reads data from an LTSpice RAW file.
-The main class object is the LTSpiceRawRead which is initialized with the filename of the RAW file to be processed.
+The main class object is the RawRead which is initialized with the filename of the RAW file to be processed.
 The object wil read the file and construct a structure of objects which can be used to access the data inside the
 RAW file.
 To understand why this is done so, in the next section follows a brief explanation of what is contained inside a RAW
@@ -150,19 +150,19 @@ header. RawReader supports both Normal and Fast Access formats
 Classes Defined
 ===============
 
-The .RAW file is read during the construction (constructor method) of an `LTSpiceRawRead` object. All traces on the RAW
+The .RAW file is read during the construction (constructor method) of an `RawRead` object. All traces on the RAW
 file are uploaded into memory.
 
-The LTSpiceRawRead class then has all the methods that allow the user to access the Axis and Trace Values. If there is
-any stepped data (.STEP primitives), the LTSpiceRawRead class will try to load the log information from the same
+The RawRead class then has all the methods that allow the user to access the Axis and Trace Values. If there is
+any stepped data (.STEP primitives), the RawRead class will try to load the log information from the same
 directory as the raw file in order to obtain the STEP information.
 
-Follows an example of the LTSpiceRawRead class usage. Information on the LTSpiceRawRead methods can be found here.
+Follows an example of the RawRead class usage. Information on the RawRead methods can be found here.
 
 Examples
 ========
 
-The example below demonstrates the usage of the LTSpiceRawRead class. It reads a .RAW file and uses the matplotlib
+The example below demonstrates the usage of the RawRead class. It reads a .RAW file and uses the matplotlib
 library to plot the results of three traces in two subplots. ::
 
     import matplotlib.pyplot as plt  # Imports the matplotlib library for plotting the results
@@ -395,7 +395,7 @@ class Axis(DataSet):
             i += 1
 
         if k != len(self.step_info):
-            raise LTSPiceReadException("The file a different number of steps than expected.\n" +
+            raise SpiceReadException("The file a different number of steps than expected.\n" +
                                        "Expecting %d got %d" % (len(self.step_offsets), k))
 
     def step_offset(self, step: int):
@@ -629,11 +629,11 @@ class DummyTrace(object):
         self.whattype = whattype
 
 
-class LTSPiceReadException(Exception):
+class SpiceReadException(Exception):
     """Custom class for exception handling"""
 
 
-class LTSpiceRawRead(object):
+class RawRead(object):
     """Class for reading LTSpice wave Files. It can read all types of Files. If stepped data is detected,
     it will also try to read the corresponding LOG file so to retrieve the stepped data.
 
@@ -846,7 +846,7 @@ class LTSpiceRawRead(object):
                         var.data[point] = float(value)
         else:
             raw_file.close()
-            raise LTSPiceReadException("Unsupported RAW File. ""%s""" % self.raw_type)
+            raise SpiceReadException("Unsupported RAW File. ""%s""" % self.raw_type)
 
         raw_file.close()
 
@@ -866,7 +866,7 @@ class LTSpiceRawRead(object):
         if "stepped" in self.raw_params["Flags"]:
             try:
                 self._load_step_information(raw_filename)
-            except LTSPiceReadException:
+            except SpiceReadException:
                 print("LOG file not found or problems happened while reading it. Auto-detecting steps")
                 if has_axis:
                     number_of_steps = 0
@@ -970,15 +970,15 @@ class LTSpiceRawRead(object):
     def _load_step_information(self, filename: Path):
         # Find the extension of the file
         if not filename.suffix == ".raw":
-            raise LTSPiceReadException("Invalid Filename. The file should end with '.raw'")
+            raise SpiceReadException("Invalid Filename. The file should end with '.raw'")
         logfile = filename.with_suffix(".log")
         try:
             encoding = detect_encoding(logfile, "Circuit:")
             log = open(logfile, 'r', errors='replace', encoding=encoding)
         except OSError:
-            raise LTSPiceReadException("Step information needs the '.log' file generated by LTSpice")
+            raise SpiceReadException("Step information needs the '.log' file generated by LTSpice")
         except UnicodeError:
-            raise LTSPiceReadException("Unable to parse log file and obtain .STEP information. Check Unicode")
+            raise SpiceReadException("Unable to parse log file and obtain .STEP information. Check Unicode")
 
         for line in log:
             if line.startswith(".step"):
@@ -1041,8 +1041,8 @@ class LTSpiceRawRead(object):
                 return range(len(self.steps))  # Returns all the steps
 
 
-# Instruction for compatibility with modification made by other users
-RawRead = LTSpiceRawRead
+# Backward compatibility naming
+LTSpiceRawRead = RawRead
 
 if __name__ == "__main__":
     import sys
