+++ PyLTSpice/RawWrite.py
@@ -2,7 +2,7 @@
 # coding=utf-8
 
 # -------------------------------------------------------------------------------
-# Name:        LTSpice_RawWrite.py
+# Name:        RawWrite.py
 # Purpose:     Create RAW Files
 #
 # Author:      Nuno Brum (nuno.brum@gmail.com)
@@ -18,7 +18,7 @@ It can be used to combine RAW files generated by different Simulation Runs
 from typing import Union
 from time import strftime
 
-from PyLTSpice.LTSpice_RawRead import DataSet, LTSpiceRawRead, Trace as TraceRead
+from PyLTSpice.RawRead import DataSet, RawRead, Trace as TraceRead
 from numpy import array, float32, zeros
 
 
@@ -58,10 +58,10 @@ class Trace(DataSet):
             self.data[:] = data[:]  # This way the dtype is kept
 
 
-class LTSpiceRawWrite(object):
+class RawWrite(object):
     """
-    This class represents the RAW data file being generated. Contrary to the the LTSpiceRawRead this class doesn't
-    support stepped data.
+    This class represents the RAW data file being generated. Contrary to the RawRead this class doesn't support stepped
+    data.
 
     """
 
@@ -135,7 +135,7 @@ class LTSpiceRawWrite(object):
         if len(self._imported_data):
             self._consolidate()
         f = open(filename, 'wb')
-        f.write("Title: * PyLTSpice LTSpice_RawWrite\n".encode(self.encoding))
+        f.write("Title: * PyLTSpice RawWrite\n".encode(self.encoding))
         f.write("Date: {}\n".format(strftime("%a %b %d %H:%M:%S %Y")).encode(self.encoding))
         f.write("Plotname: {}\n".format(self.plot_name).encode(self.encoding))
         f.write("Flags: {}\n".format(self._str_flags()).encode(self.encoding))
@@ -180,12 +180,12 @@ class LTSpiceRawWrite(object):
                 return True
         return False
 
-    def add_traces_from_raw(self, other: LTSpiceRawRead, trace_filter: Union[list, tuple, str], **kwargs):
+    def add_traces_from_raw(self, other: RawRead, trace_filter: Union[list, tuple, str], **kwargs):
         """ *(Not fully implemented)*
 
-        Merge two LTSpiceRawWrite classes together resulting in a new instance
-        :param other: an instance of the LTSpiceRawRead class where the traces are going to be copied from.
-        :type other: LTSpiceRawRead
+        Merge two RawWrite classes together resulting in a new instance
+        :param other: an instance of the RawRead class where the traces are going to be copied from.
+        :type other: RawRead
         :param trace_filter: A list of signals that should be imported into the new file
         :type trace_filter: list, Tuple, or just a string for one trace
 
@@ -351,10 +351,10 @@ def tobytes_for_trace(trace: Trace):
 
 if __name__ == '__main__':
     import numpy as np
-    from LTSpice_RawRead import LTSpiceRawRead
+    from RawRead import RawRead
 
     def test_readme_snippet():
-        LW = LTSpiceRawWrite(fastacces=False)
+        LW = RawWrite(fastacces=False)
         tx = Trace('time', np.arange(0.0, 3e-3, 997E-11))
         vy = Trace('N001', np.sin(2 * np.pi * tx.data * 10000))
         vz = Trace('N002', np.cos(2 * np.pi * tx.data * 9970))
@@ -378,7 +378,7 @@ if __name__ == '__main__':
                     break
         if raw_type == 'transient' and wave_size > 0:
             data = np.genfromtxt(f, dtype='float,float', delimiter=',', max_rows=wave_size)
-            LW = LTSpiceRawWrite()
+            LW = RawWrite()
             LW.add_trace(Trace('time', [x[0] for x in data]))
             LW.add_trace(Trace('Ampl', [x[1] for x in data]))
             LW.save("teste_trc.raw")
@@ -386,7 +386,7 @@ if __name__ == '__main__':
 
 
     def test_axis_sync():  # Test axis sync
-        LW = LTSpiceRawWrite()
+        LW = RawWrite()
         tx = Trace('time', np.arange(0.0, 3e-3, 997E-11))
         vy = Trace('N001', np.sin(2 * np.pi * tx.data * 10000))
         vz = Trace('N002', np.cos(2 * np.pi * tx.data * 9970))
@@ -394,7 +394,7 @@ if __name__ == '__main__':
         LW.add_trace(vy)
         LW.add_trace(vz)
         LW.save("teste_w.raw")
-        LR = LTSpiceRawRead("..\\test_files\\testfile.raw")
+        LR = RawRead("..\\test_files\\testfile.raw")
         LW.add_traces_from_raw(LR, ('V(out)',), force_axis_alignment=True)
         LW.save("merge.raw")
         test = """
@@ -416,9 +416,9 @@ if __name__ == '__main__':
         """
 
     def test_write_ac():
-        LW = LTSpiceRawWrite()
-        LR = LTSpiceRawRead("..\\tests\\PI_Filter.raw")
-        LR1 = LTSpiceRawRead("..\\tests\\PI_Filter_resampled.raw")
+        LW = RawWrite()
+        LR = RawRead("..\\tests\\PI_Filter.raw")
+        LR1 = RawRead("..\\tests\\PI_Filter_resampled.raw")
         LW.add_traces_from_raw(LR, ('V(N002)',))
         LW.add_traces_from_raw(LR1, 'V(N002)', rename_format='N002_resampled', force_axis_alignment=True)
         LW.flag_fastaccess = False
@@ -427,8 +427,8 @@ if __name__ == '__main__':
         LW.save("..\\tests\\PI_filter_rewritten_fast.raw")
 
     def test_write_tran():
-        LR = LTSpiceRawRead("..\\tests\\TRAN - STEP.raw")
-        LW = LTSpiceRawWrite()
+        LR = RawRead("..\\tests\\TRAN - STEP.raw")
+        LW = RawWrite()
         LW.add_traces_from_raw(LR, ('V(out)', 'I(C1)'))
         LW.flag_fastaccess = False
         LW.save("..\\tests\\TRAN - STEP0_normal.raw")
@@ -436,7 +436,7 @@ if __name__ == '__main__':
         LW.save("..\\tests\\TRAN - STEP0_fast.raw")
 
     def test_combine_tran():
-        LW = LTSpiceRawWrite()
+        LW = RawWrite()
         for tag, raw in (
             ("AD820_15", "../tests/Batch_Test_AD820_15.raw"),
             # ("AD820_10", "../tests/Batch_Test_AD820_10.raw"),
@@ -445,7 +445,7 @@ if __name__ == '__main__':
             # ("AD820_5", "../tests/Batch_Test_AD820_5.raw"),
             # ("AD712_5", "../tests/Batch_Test_AD712_5.raw"),
         ):
-            LR = LTSpiceRawRead(raw)
+            LR = RawRead(raw)
             LW.add_traces_from_raw(LR, ("V(out)", "I(R1)"), rename_format="{}_{tag}", tag=tag, force_axis_alignment=True)
         LW.flag_fastaccess = False
         LW.save("../tests/Batch_Test_Combine.raw")
