+++ setup.py
@@ -9,10 +9,10 @@ setuptools.setup(
     scripts=['PyLTSpice/__init__.py',
              'PyLTSpice/detect_encoding.py',
              'PyLTSpice/Histogram.py',
-             'PyLTSpice/LTSpice_RawRead.py',
-             'PyLTSpice/LTSpice_RawWrite.py',
-             'PyLTSpice/LTSpice_SemiDevOpReader.py',
-             'PyLTSpice/LTSpiceBatch.py',
+             'PyLTSpice/RawRead.py',
+             'PyLTSpice/RawWrite.py',
+             'PyLTSpice/SemiDevOpReader.py',
+             'PyLTSpice/SpiceBatch.py',
              'PyLTSpice/LTSteps.py',
              'PyLTSpice/sim_stepping.py',
              'PyLTSpice/SpiceEditor.py',
